Fleshed out README and settings.py, and added collectstatic and sqlite folders.
The collectstatic and sqlite folders have had their svn:ignore property set so
as to ignore *all* files contained therein. This is appropriate because they
contain site-specific data.

The README has been updated with a brief explanation of the structure of the
website.

The settings file has been updated such that sqlite is the default database.


git-svn-id: https://subversion.assembla.com/svn/elts@4 3fefdeda-3626-427b-91b0-c7b7be374f37