Add ability to create and read `Tag` db table
The ability to ineract with items from the `Tag` database table has been added.
It is now possible to create new tags and read existing tags. It is not yet
possible to update or delete existing tags.

Logic for viewing items has also been updated. If viewing a page for an item
that does not exist (say, `/item/9001`), the logic in `views.py` will no longer
throw an exception. Instead, a proper web page will be produced, informing the
user that the requested item does not exist.


git-svn-id: https://subversion.assembla.com/svn/elts@52 3fefdeda-3626-427b-91b0-c7b7be374f37