Add ability to update existing items and tags
Views have been updated with logic necessary to update and items and tags.
Unfortunately, the forms used to update items and tags are not currently
pre-filled with existing data, and in case of update failure, nothing is
returned to the user. Also, the DELETE operator still has not been implemented,
and proper object validation is *still* not being done before saving objects.

The HTML templates have been updated so links to the update forms are in place.

Comments have been trimmed down to make code easier on the eyes.


git-svn-id: https://subversion.assembla.com/svn/elts@57 3fefdeda-3626-427b-91b0-c7b7be374f37