Refactor `Note` model
The Note model is now an abstract class. Note defines the common fields
note_text, note_date, and note_author, and it's children (ItemNote, PersonNote,
ReservationNote, and LendNote) simply add foreign keys.

This was done because the existing many-to-many relationships between Note and X
(e.g. Item) was an ill fit. Notes are always attached to a *single* other row,
not many. The relationship is now one-to-many.


git-svn-id: https://subversion.assembla.com/svn/elts@76 3fefdeda-3626-427b-91b0-c7b7be374f37