Merge Lend and Reservation tables.
The Lend and Reservation tables have been merged. All tracking of when an item
should be lent out and hwen it actually does go out is now in one place.

The ``due_back_date`` attribute has been removed from the ``Item`` table. This
information should be stored in the ``Lend`` table. Additionally, an
``is_lendable`` field has been added to the ``Item`` table to denote whether it
is available for normal lending.

An ``is_complaint`` attribute has been added to the ``LendNote`` table. This
allows item notes to be marked as a complaint or not a complaint in a very
simple way.

Templates and forms have been adjusted to allow for these changes.


git-svn-id: https://subversion.assembla.com/svn/elts@83 3fefdeda-3626-427b-91b0-c7b7be374f37