Rename fields for model `Lend`
The following field renames were performed for model `Lend`:

* out_reservation => due_out
* back_reservation => due_back
* out_actual => out
* back_actual => back

Code was updated as necessary to accomodate this change.

Also DRY up the `elts/lend.html` template. The names of several column names are
now dynamically generated from the `verbose_namme` of a model's field names
instead of being hard coded. In other words, this:

    <th>Item</th>

has been replaced by this:

    <th>{{ lends|first|label:"item_id"|capfirst }}</th>

The template tag `label` was written to accomodate this change. This template
tag does not have any unit tests, and it does not fail gracefully. It should be
made more robust, a doctest should be written, and the filter should be used
more extensively (many templates have hard-coded field labels).