Update page titles.
Make the titles of several pages more consistent.

The page titles, as they stand, are currently awkward and rather useless. It
might be better if, for example, the very top header on the page is used as a
breakcrumb nav section, linking to each portion of the current URL (RESTful URL
design makes this possible). The more friendly and informative page title could
then reside with the remainder of the content. For example, this might go at the
very top of a page:

    <ol>
        <li><a>Item</a></li>
        <li><a>{{ item.id }}</a></li>
        <li><a>Update Form</a></li>
    </ol>

And this would go at the top of the content just following it:

    <h1>{{ item.name }}</h1>