Use django-tables2 to render table of lends
The changes in this commit are the bare necessities for creating an HTML table
from a Django data set. Some of the changes are nice: defining how Django data
sets can be transformed into HTML tables with a file named "tables.py" is
familiar and seemingly elegant. Other changes are not so nice:
TEMPLATE_CONTEXT_PROCESSORS has been added to settings.py, the generated HTML is
kinda ugly, and this seems like a heavyweight solution to a seemingly simple
problem.