Make calendar show past and future months
Make the `elts.views.calendar` function fetch the URL parameter 'month_offset'
and use it to determine which month should be displayed. The
`elts.views._increment_month()` and `elts.views._convert_to_int()` functions
help in this endeavour.

Rename the template parameter 'today' to 'target_date'.

Update the calendar template with links for viewing the previous or next month.