Move hack from test_views.py to factories.py
The python 2 JSON encoder has a known issue with UTF-8 characters, and this
limitation is compensated for by creating sub-par random data. Move this hack
from test_views.py to factories.py. If we ever upgrade to python 3, we can now
remove the hack from a single place in the source code, instead of multiple
locations.