Allow HTTP GET /category/<id>/ to succeed.
Make it possible for a client to issue an HTTP GET request to /category/<id>/
URLs. In other words, a client can now fetch information about a particular
category. This change makes the category* views more in line with other families
of views, such as the item* and tag* views.

Update various views to point to issue HTTP redirects to this new view, and also
update tests to match.